+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # [Releases](https://github.com/Tracktor/design-system/releases)
 
-## v4.28.0
+## v4.29.0
 
 ### Feat 
-- add customizable participant name formatting to conversation components
+- Add new kanban variant
+- Improve drag and drop performance for kanban